Properly escape backticks for XSS solutions
The backticks were not being rendered properly in the solution, leading to invalid code when copy-pasting from it. This PR updates the formatting to fix this issue by using double-backticks to enclose the relevant code, so that single backticks are rendered properly. I have not exhaustively checked if I caught all instances of this issue.